Created directory and page items
Created and linked all of the pages to the main app.js file. Currently the page only shows white with no errors when I try to add in the navigation links on the main app.js page.

When the links are removed the app.js file is correctly importing content from the header and footer components.

Next steps once the navigation is fixed will be styling the pages.